issue 136  delete for unsigned

+++ src/components/CardMovieEditable.vue
@@ -1,6 +1,13 @@
 <template>
   <div
-    class="movie-card rounded-lg text-gray-200 mb-4 flex flex-col overflow-hidden"
+    class="
+      movie-card
+      rounded-lg
+      text-gray-200
+      mb-4
+      flex flex-col
+      overflow-hidden
+    "
     :class="excludeMovie ? 'bg-gray-900 movie-card--exclude' : 'bg-gray-800'"
     @click="showDetails = !showDetails"
   >
@@ -209,9 +216,11 @@ export default class CardMovieEditable extends Vue {
   }
 
   deleteMovie (): void {
-    db.collection(this.$store.getters.getCurrentUserDocumentId)
-      .doc(this.movie.documentId)
-      .delete();
+    if (this.isSignedIn) {
+      db.collection(this.$store.getters.getCurrentUserDocumentId)
+        .doc(this.movie.documentId)
+        .delete();
+    }
     this.$store.commit("deleteMovieFromList", this.movie);
   }
